WELD-2698 Remove MR JAR setup, keep JDK 11 variant of WeldDefaultProxyServices.
